+++ src/executor/monitor/monitor.cpp
@@ -9,6 +9,8 @@
 #include "platform_interrupt_controller.hpp"
 #include "utility/crc32.hpp"
 
+#include "xpseudo_asm.h"
+
 using namespace bmboot;
 using namespace bmboot::internal;
 
@@ -21,6 +23,9 @@ extern "C" int main()
     auto& ipc_block = (volatile IpcBlock &) getIpcBlock();
     volatile const auto& inbox = ipc_block.manager_to_executor;
     volatile auto& outbox = ipc_block.executor_to_manager;
+
+    // This is normally set by the firmware... plot twist -- we're the firmware now.
+    mtcp(CNTFRQ_EL0, inbox.cntfrq);
 
     platform::setupInterrupts();